Move story legend to separate component
